backend/requirements.txt
fastapi
uvicorn
pydantic
bcrypt
pyjwt
